Fix padding calculation
Suppose *header_align* = 4, *header_size* = 8, and *elem_align* = 8.  The
current code will align to 4 bytes (*header_align*) and add 4 bytes of padding
after the header (*elem_align* - *header_align*).

But the allocater **may** return a buffer that happens to be 8-byte aligned, in
which case the end of the header is also 8-byte aligned, and the correct
padding is 0 bytes, not 4.

This patch fixes this case by aligning to max(*header_align*, *elem_align*)
and adding padding based on *header_size* mod *alignment*.