Don't allocate a buffer with capacity 0.
This fixes a memory leak in the following code:

    let mut v = ThinVec::new();
    v.push(1);
    v.pop();
    v.shrink_to_fit();

This leaks because `shrink_to_fit` creates a new allocation with
capacity zero.  When the vector is later dropped or resized,
`reallocate` or `deallocate` assume that it has no allocation, because
the capacity is zero.

(This leak can also be triggered by `ThinVec::with_capacity(0)`.)

The fix attacks the problem from both directions: It avoids allocating a
buffer with capacity 0 to begin with, and it also makes the checks in
`deallocate` and `reallocate` robust against this situation.