0.2.3 release
This includes a memory safety fix for ThinVecs containing types with higher alignment requirements than pointers. These types are already disallowed by the gecko_ffi feature, so users of that feature were unaffected. Higher alignment than a pointer is very rare and generally only possible with explicit annotations. The code was very blatantly broken for these over-aligned types, so it's unlikely anyone was actually using them with ThinVec in the wild.